rename file to be more in line with other switches
