+++ src/services/ui/src/features/package-actions/lib/index.ts
@@ -6,7 +6,7 @@ import { buildActionUrl, useOriginPath } from "@/utils";
 import { Route, useAlertContext, useNavigate } from "@/components";
 import { FieldValues } from "react-hook-form";
 import { getContentFor } from "@/features/package-actions/lib/contentSwitch";
-import { successCheckSwitch } from "./correctStatusSwitch";
+import { successCheckSwitch } from "./successCheckSwitch";
 import { documentPoller } from "@/utils/Poller/seaStatusPoller";
 import { stripQueryStringFromURL } from "@/utils/stripQueryString";
 import { SPA_ID_REGEX } from "@/consts";
